@@ -1,136 +1,97 @@
 {
-	"name": "adminpage",
-	"version": "3.0.0",
-	"description": "admin page for carrectly",
-	"engines": {
-		"node": "12.x"
-	},
-	"main": "index.js",
-	"scripts": {
-		"build-client": "webpack",
-		"build-client-watch": "webpack -w",
-		"deploy": "script/deploy",
-		"heroku-token": "script/encrypt-heroku-auth-token.js",
-		"lint": "eslint ./ --ignore-path .gitignore",
-		"lint-fix": "npm run lint -- --fix",
-		"prepare": "if [ -d .git ]; then npm-merge-driver install; fi",
-		"prettify": "prettier --write \"**/*.{js,jsx,json,css,scss,md}\"",
-		"postinstall": "touch secrets.js",
-		"seed": "node server/db/seed.js",
-		"start": "node server/index.js",
-		"start-dev": "NODE_ENV='development' npm run build-client-watch & NODE_ENV='development' npm run start-server",
-		"start-server": "nodemon server -e html,js,scss --ignore public --ignore client"
-	},
-	"husky": {
-		"hooks": {
-			"pre-commit": "lint-staged"
-		}
-	},
-	"lint-staged": {
-		"*.{js,jsx}": [
-			"prettier --write",
-			"npm run lint-fix",
-			"git add"
-		],
-		"*.{css,scss,json,md}": [
-			"prettier --write",
-			"git add"
-		]
-	},
-	"author": "Andre Birkus",
-	"license": "MIT",
-	"dependencies": {
-		"@ant-design/icons": "^4.2.1",
-		"@fortawesome/fontawesome-svg-core": "^1.2.28",
-		"@fortawesome/free-solid-svg-icons": "^5.13.0",
-		"@fortawesome/react-fontawesome": "^0.1.11",
-		"antd": "^4.3.4",
-		"axios": "^0.21.1",
-		"compression": "^1.7.3",
-		"connect-session-sequelize": "^6.0.0",
-		"dom-parser": "^0.1.6",
-		"express": "^4.16.4",
-		"express-session": "^1.15.1",
-		"faker": "^4.1.0",
-		"gmail-api-parse-message": "^2.1.0",
-		"googleapis": "^39.2.0",
-		"history": "^4.9.0",
-		"html-react-parser": "^0.10.0",
-		"moment": "^2.24.0",
-		"morgan": "^1.9.1",
-		"open": "^7.0.0",
-		"openurl": "^1.1.1",
-		"passport": "^0.4.0",
-		"passport-google-oauth": "^2.0.0",
-		"pg": "^7.9.0",
-		"pg-hstore": "^2.3.2",
-		"pg-native": "^3.0.0",
-		"prop-types": "^15.7.2",
-		"react": "^16.8.6",
-		"react-dom": "^16.8.6",
-		"react-google-calendar-api": "^1.0.11",
-		"react-highlight-words": "^0.16.0",
-		"react-newline-to-break": "^1.0.6",
-		"react-redux": "^7.0.1",
-		"react-render-html": "^0.6.0",
-		"react-router-dom": "^5.0.0",
-		"redux": "^4.0.1",
-		"redux-logger": "^3.0.6",
-		"redux-thunk": "^2.3.0",
-		"sequelize": "^5.3.1",
-		"server-destroy": "^1.0.1",
-		"socket.io": "^2.4.0",
-		"square": "^8.1.1",
-		"travis-cli": "^1.0.9",
-		"window": "^4.2.6"
-	},
-	"devDependencies": {
-		"@babel/core": "^7.4.3",
-		"@babel/plugin-proposal-class-properties": "7.4.0",
-		"@babel/plugin-proposal-decorators": "7.4.0",
-		"@babel/plugin-proposal-export-namespace-from": "7.2.0",
-		"@babel/plugin-proposal-function-sent": "7.2.0",
-		"@babel/plugin-proposal-json-strings": "7.2.0",
-		"@babel/plugin-proposal-numeric-separator": "7.2.0",
-		"@babel/plugin-proposal-throw-expressions": "7.2.0",
-		"@babel/plugin-syntax-dynamic-import": "7.2.0",
-		"@babel/plugin-syntax-import-meta": "7.2.0",
-		"@babel/polyfill": "^7.4.3",
-		"@babel/preset-env": "^7.4.3",
-		"@babel/preset-react": "^7.0.0",
-		"@babel/register": "^7.4.0",
-		"axios-mock-adapter": "^1.16.0",
-		"babel-eslint": "^10.0.1",
-		"babel-loader": "^8.0.5",
-		"chai": "^4.2.0",
-		"css-loader": "^2.1.1",
-		"enzyme": "^3.9.0",
-		"enzyme-adapter-react-16": "^1.12.1",
-		"eslint": "^5.16.0",
-		"eslint-config-fullstack": "^6.0.0",
-		"eslint-config-prettier": "^4.1.0",
-		"eslint-plugin-react": "^7.12.4",
-		"fibers": "^5.0.0",
-		"git-url-parse": "^11.1.2",
-		"husky": "^1.3.1",
-		"lint-staged": "^8.1.5",
-		"mocha": "^6.1.4",
-		"node-sass": "^5.0.0",
-		"nodemon": "^1.18.3",
-		"npm-merge-driver": "^2.3.6",
-		"open": "^7.0.0",
-		"prettier": "^1.19.1",
-		"react-test-renderer": "^16.4.2",
-		"redux-devtools-extension": "^2.13.5",
-		"redux-mock-store": "^1.5.3",
-		"sass": "^1.32.8",
-		"sass-loader": "^7.2.0",
-		"server-destroy": "^1.0.1",
-		"simple-git": "^1.121.0",
-		"style-loader": "^0.21.0",
-		"supertest": "^4.0.2",
-		"webpack": "^4.16.4",
-		"webpack-cli": "^3.3.11",
-		"yaml": "^1.6.0"
-	}
+  "name": "adminpage",
+  "version": "4.0.0",
+  "description": "administrative website for car service business",
+  "engines": {
+    "node": "16.x"
+  },
+  "main": "index.js",
+  "scripts": {
+    "build-client": "webpack",
+    "build-client-watch": "webpack -w",
+    "deploy": "script/deploy",
+    "heroku-token": "script/encrypt-heroku-auth-token.js",
+    "lint": "eslint ./ --ignore-path .gitignore",
+    "lint-fix": "npm run lint -- --fix",
+    "prepare": "if [ -d .git ]; then npm-merge-driver install; fi",
+    "prettify": "prettier --write \"**/*.{js,jsx,json,css,scss,md}\"",
+    "postinstall": "touch secrets.js",
+    "seed": "node server/db/seed.js",
+    "start": "node server/index.js",
+    "start-dev": "NODE_ENV='development' npm run build-client-watch & NODE_ENV='development' npm run start-server",
+    "start-server": "nodemon server -e html,js,scss --ignore public --ignore client"
+  },
+  "author": "Andre Birkus",
+  "license": "MIT",
+  "dependencies": {
+    "@ant-design/icons": "^4.7.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "antd": "^4.17.3",
+    "axios": "^0.24.0",
+    "compression": "^1.7.4",
+    "connect-session-sequelize": "^7.1.2",
+    "dom-parser": "^0.1.6",
+    "express": "^4.17.1",
+    "express-session": "^1.17.2",
+    "faker": "^5.5.3",
+    "gmail-api-parse-message": "^2.1.2",
+    "googleapis": "^92.0.0",
+    "history": "^5.1.0",
+    "html-react-parser": "^1.4.2",
+    "moment": "^2.29.1",
+    "morgan": "^1.10.0",
+    "openurl": "^1.1.1",
+    "passport": "^0.5.0",
+    "passport-google-oauth": "^2.0.0",
+    "pg": "^8.7.1",
+    "pg-hstore": "^2.3.4",
+    "pg-native": "^3.0.0",
+    "prop-types": "^15.7.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-highlight-words": "^0.17.0",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.0.2",
+    "redux": "^4.1.2",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.4.1",
+    "sequelize": "^6.11.0",
+    "server-destroy": "^1.0.1",
+    "socket.io": "^4.4.0",
+    "square": "^16.0.0",
+    "travis-cli": "^1.0.9",
+    "window": "^4.2.7"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-react": "^7.16.0",
+    "@webpack-cli/generators": "^2.4.1",
+    "babel-loader": "^8.2.3",
+    "css-loader": "^6.5.1",
+    "eslint": "^8.4.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.27.1",
+    "file-loader": "^6.2.0",
+    "git-url-parse": "^11.6.0",
+    "html-webpack-plugin": "^5.5.0",
+    "lint-staged": "^12.1.2",
+    "mini-css-extract-plugin": "^2.4.4",
+    "node-sass": "^7.0.0",
+    "nodemon": "^2.0.15",
+    "npm-merge-driver": "^2.3.6",
+    "prettier": "^2.5.1",
+    "redux-devtools-extension": "^2.13.9",
+    "sass": "^1.44.0",
+    "sass-loader": "^12.4.0",
+    "server-destroy": "^1.0.1",
+    "simple-git": "^2.48.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.6.0"
+  }
 }
