@@ -1,19 +1,19 @@
-import React, {useState, useEffect} from 'react'
-import {useDispatch, useSelector} from 'react-redux'
-import {getContactsThunk} from '../../store/contacts'
+import React, { useState, useEffect } from 'react'
+import { useDispatch } from 'react-redux'
+import { getContactsThunk } from '../../store/contacts'
 import AndtDCustomerTable from './AntDCustomersTable'
 
 const AllCustomers = () => {
-	const [loading, setLoading] = useState(false)
-	const dispatch = useDispatch()
+  const [loading, setLoading] = useState(false)
+  const dispatch = useDispatch()
 
-	useEffect(() => {
-		setLoading(true)
-		dispatch(getContactsThunk())
-		setLoading(false)
-	}, [])
+  useEffect(() => {
+    setLoading(true)
+    dispatch(getContactsThunk())
+    setLoading(false)
+  }, [])
 
-	return <AndtDCustomerTable loading={loading} />
+  return <AndtDCustomerTable loading={loading} />
 }
 
 export default AllCustomers
