@@ -1,24 +1,20 @@
-import React, {Component} from 'react'
-import {withRouter} from 'react-router-dom'
-import {connect} from 'react-redux'
-import {Table} from 'antd'
+import React from 'react'
+import { Table } from 'antd'
 import columns from '../Table/HomeTableColumns'
 
-class TableOrdersByStatus extends Component {
-	render() {
-		const array = this.props.ordersArray || []
+const TableOrdersByStatus = (props) => {
+  const array = props.ordersArray || []
 
-		return (
-			<Table
-				scroll={{x: 'max-content'}}
-				columns={columns}
-				dataSource={array}
-				pagination={false}
-				size='small'
-				rowKey='hash'
-			/>
-		)
-	}
+  return (
+    <Table
+      scroll={{ x: 'max-content' }}
+      columns={columns}
+      dataSource={array}
+      pagination={false}
+      size="small"
+      rowKey="hash"
+    />
+  )
 }
 
-export default withRouter(connect(null, null)(TableOrdersByStatus))
+export default TableOrdersByStatus
