@@ -1,15 +1,15 @@
-import React, {Component} from 'react'
-import {useSelector} from 'react-redux'
-import renderHTML from 'react-render-html'
+import React from 'react'
+import { useSelector } from 'react-redux'
 
 const SingleEmail = () => {
-	const message = useSelector(state => state.singleemail)
+  const message = useSelector((state) => state.singleemail)
 
-	if (message.length) {
-		return <div>{renderHTML(message)}</div>
-	} else {
-		return <div>Click on the inbox subject to preview</div>
-	}
+  if (message.length) {
+    // return <div>{renderHTML(message)}</div>
+    return <div></div>
+  } else {
+    return <div>Click on the inbox subject to preview</div>
+  }
 }
 
 export default SingleEmail
