@@ -1,84 +1,79 @@
 /* eslint-disable react/display-name */
-import React, {useState, useEffect, useRef} from 'react'
+import React from 'react'
 import Highlighter from 'react-highlight-words'
-import {SearchOutlined} from '@ant-design/icons'
-import {Table, Input, Button, Space} from 'antd'
+import { SearchOutlined } from '@ant-design/icons'
+import { Input, Button, Space } from 'antd'
 
 const getColumnSearchProps = (
-	dataIndex,
-	searchInput,
-	searchText,
-	searchedColumn,
-	handleSearch,
-	handleReset
+  dataIndex,
+  searchInput,
+  searchText,
+  searchedColumn,
+  handleSearch,
+  handleReset
 ) => ({
-	filterDropdown: ({
-		setSelectedKeys,
-		selectedKeys,
-		confirm,
-		clearFilters,
-	}) => {
-		return (
-			<div style={{padding: 8}}>
-				<Input
-					ref={searchInput}
-					placeholder={`Search ${dataIndex}`}
-					value={selectedKeys[0]}
-					onChange={e =>
-						setSelectedKeys(e.target.value ? [e.target.value] : [])
-					}
-					onPressEnter={() =>
-						handleSearch(selectedKeys, confirm, dataIndex)
-					}
-					style={{width: 188, marginBottom: 8, display: 'block'}}
-				/>
-				<Space>
-					<Button
-						type='primary'
-						onClick={() =>
-							handleSearch(selectedKeys, confirm, dataIndex)
-						}
-						icon={<SearchOutlined />}
-						size='small'
-						style={{width: 90}}>
-						Search
-					</Button>
-					<Button
-						onClick={() => handleReset(clearFilters)}
-						size='small'
-						style={{width: 90}}>
-						Reset
-					</Button>
-				</Space>
-			</div>
-		)
-	},
-	filterIcon: filtered => (
-		<SearchOutlined style={{color: filtered ? '#1890ff' : undefined}} />
-	),
-	onFilter: (value, record) =>
-		record[dataIndex]
-			? record[dataIndex]
-					.toString()
-					.toLowerCase()
-					.includes(value.toLowerCase())
-			: '',
-	onFilterDropdownVisibleChange: visible => {
-		if (visible) {
-			setTimeout(() => searchInput.current.select(), 100)
-		}
-	},
-	render: text =>
-		searchedColumn === dataIndex ? (
-			<Highlighter
-				highlightStyle={{backgroundColor: '#ffc069', padding: 0}}
-				searchWords={[searchText]}
-				autoEscape
-				textToHighlight={text ? text.toString() : ''}
-			/>
-		) : (
-			text
-		),
+  filterDropdown: ({
+    setSelectedKeys,
+    selectedKeys,
+    confirm,
+    clearFilters,
+  }) => {
+    return (
+      <div style={{ padding: 8 }}>
+        <Input
+          ref={searchInput}
+          placeholder={`Search ${dataIndex}`}
+          value={selectedKeys[0]}
+          onChange={(e) =>
+            setSelectedKeys(e.target.value ? [e.target.value] : [])
+          }
+          onPressEnter={() => handleSearch(selectedKeys, confirm, dataIndex)}
+          style={{ width: 188, marginBottom: 8, display: 'block' }}
+        />
+        <Space>
+          <Button
+            type="primary"
+            onClick={() => handleSearch(selectedKeys, confirm, dataIndex)}
+            icon={<SearchOutlined />}
+            size="small"
+            style={{ width: 90 }}
+          >
+            Search
+          </Button>
+          <Button
+            onClick={() => handleReset(clearFilters)}
+            size="small"
+            style={{ width: 90 }}
+          >
+            Reset
+          </Button>
+        </Space>
+      </div>
+    )
+  },
+  filterIcon: (filtered) => (
+    <SearchOutlined style={{ color: filtered ? '#1890ff' : undefined }} />
+  ),
+  onFilter: (value, record) =>
+    record[dataIndex]
+      ? record[dataIndex].toString().toLowerCase().includes(value.toLowerCase())
+      : '',
+  onFilterDropdownVisibleChange: (visible) => {
+    if (visible) {
+      setTimeout(() => searchInput.current.select(), 100)
+    }
+  },
+  render: (text) =>
+    searchedColumn === dataIndex ? (
+      <Highlighter
+        highlightStyle={{ backgroundColor: '#ffc069', padding: 0 }}
+        searchWords={[searchText]}
+        autoEscape
+        textToHighlight={text ? text.toString() : ''}
+      />
+    ) : (
+      text
+    ),
 })
 
 export default getColumnSearchProps
