@@ -1,18 +1,19 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
+import { Provider } from 'react-redux'
+import { BrowserRouter } from 'react-router-dom'
 import App from './app'
 import store from './store'
-import history from './history'
 import 'antd/dist/antd.css'
 import './components/styles/styles.scss'
 
 ReactDOM.render(
-	<Provider store={store}>
-		<Router history={history}>
-			<App />
-		</Router>
-	</Provider>,
-	document.getElementById('app')
+  <Provider store={store}>
+    {/* <Router history={history}> */}
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+    {/* </Router> */}
+  </Provider>,
+  document.getElementById('app')
 )
