@@ -6,16 +6,12 @@ const session = require('express-session')
 const passport = require('passport')
 const SequelizeStore = require('connect-session-sequelize')(session.Store)
 const db = require('./db')
-const sessionStore = new SequelizeStore({db})
+const sessionStore = new SequelizeStore({ db })
 const PORT = process.env.PORT || 1337
 const app = express()
 const socketio = require('socket.io')
 
 module.exports = app
-
-if (process.env.NODE_ENV === 'test') {
-	after('close the session store', () => sessionStore.stopExpiringSessions())
-}
 
 if (process.env.NODE_ENV !== 'production') require('../secrets')
 
@@ -23,103 +19,103 @@ if (process.env.NODE_ENV !== 'production') require('../secrets')
 passport.serializeUser((user, done) => done(null, user.id))
 
 passport.deserializeUser(async (id, done) => {
-	try {
-		const user = await db.models.user.findByPk(id)
-		done(null, user)
-	} catch (err) {
-		done(err)
-	}
+  try {
+    const user = await db.models.user.findByPk(id)
+    done(null, user)
+  } catch (err) {
+    done(err)
+  }
 })
 
 const createApp = () => {
-	// logging middleware
-	app.use(morgan('dev'))
+  // logging middleware
+  app.use(morgan('dev'))
 
-	// body parsing middleware
-	app.use(express.json())
-	app.use(express.urlencoded({extended: true}))
+  // body parsing middleware
+  app.use(express.json())
+  app.use(express.urlencoded({ extended: true }))
 
-	// compression middleware
-	app.use(compression())
+  // compression middleware
+  app.use(compression())
 
-	// session middleware with passport
-	app.use(
-		session({
-			secret: process.env.SESSION_SECRET || 'my best friend is Cody',
-			store: sessionStore,
-			resave: false,
-			saveUninitialized: true,
-		})
-	)
-	app.use(passport.initialize())
-	app.use(passport.session())
+  // session middleware with passport
+  app.use(
+    session({
+      secret: process.env.SESSION_SECRET || 'my best friend is Cody',
+      store: sessionStore,
+      resave: false,
+      saveUninitialized: true,
+    })
+  )
+  app.use(passport.initialize())
+  app.use(passport.session())
 
-	// static file-serving middleware
-	app.use(express.static(path.join(__dirname, '..', 'public')))
-	// auth and api routes
-	app.use('/auth', require('./auth'))
-	app.use('/api', require('./api'))
-	app.use('/square', require('./square'))
-	app.use('/wpbookings', require('./wpbookings'))
+  // static file-serving middleware
+  app.use(express.static(path.join(__dirname, '..', 'public')))
+  // auth and api routes
+  app.use('/auth', require('./auth'))
+  app.use('/api', require('./api'))
+  app.use('/square', require('./square'))
+  app.use('/wpbookings', require('./wpbookings'))
 
-	// any remaining requests with an extension (.js, .css, etc.) send 404
-	app.use((req, res, next) => {
-		if (path.extname(req.path).length) {
-			const err = new Error('Not found')
-			err.status = 404
-			next(err)
-		} else {
-			next()
-		}
-	})
+  // any remaining requests with an extension (.js, .css, etc.) send 404
+  app.use((req, res, next) => {
+    if (path.extname(req.path).length) {
+      const err = new Error('Not found')
+      err.status = 404
+      next(err)
+    } else {
+      next()
+    }
+  })
 
-	// sends index.html
-	app.use('*', (req, res) => {
-		res.sendFile(path.join(__dirname, '..', 'public/index.html'))
-	})
+  // sends index.html
+  app.use('*', (req, res) => {
+    res.sendFile(path.join(__dirname, '..', 'public/index.html'))
+  })
 
-	// error handling endware
-	app.use((err, req, res, next) => {
-		console.error(err)
-		console.error(err.stack)
-		if (err.errors) {
-			res.status(err.statusCode).send(err.errors)
-		} else {
-			res.status(err.status || 500).send(
-				err.message || 'Internal server error.'
-			)
-		}
-	})
+  // error handling endware
+  app.use((err, req, res, next) => {
+    console.error(err)
+    console.error(err.stack)
+    if (err.errors) {
+      res.status(err.statusCode).send(err.errors)
+    } else {
+      res
+        .status(err.status || 500)
+        .send(err.message || 'Internal server error.')
+    }
+  })
 }
 
 const startListening = () => {
-	// start listening (and create a 'server' object representing our server)
-	const server = app.listen(PORT, () =>
-		console.log(`Mixing it up on port ${PORT}`)
-	)
+  // start listening (and create a 'server' object representing our server)
+  const server = app.listen(PORT, () =>
+    console.log(`Mixing it up on port ${PORT}`)
+  )
 
-	const io = socketio(server)
-	require('./socket')(io)
+  const io = socketio(server)
+  require('./socket')(io)
 }
 
 const syncDb = () => db.sync()
 
 async function bootApp() {
-	try {
-		await sessionStore.sync()
-		await syncDb()
-		await createApp()
-		await startListening()
-	} catch (err) {
-		console.log(err)
-	}
+  try {
+    await sessionStore.sync()
+    await syncDb()
+    await createApp()
+    await startListening()
+  } catch (err) {
+    console.log(err)
+  }
 }
 // This evaluates as true when this file is run directly from the command line,
 // i.e. when we say 'node server/index.js' (or 'nodemon server/index.js', or 'nodemon server', etc)
 // It will evaluate false when this module is required by another module - for example,
 // if we wanted to require our app in a test spec
 if (require.main === module) {
-	bootApp()
+  bootApp()
 } else {
-	createApp()
+  createApp()
 }
