@@ -1,50 +1,44 @@
 const isDev = process.env.NODE_ENV === 'development'
 
 module.exports = {
-	mode: isDev ? 'development' : 'production',
-	entry: [
-		'@babel/polyfill', // enables async-await
-		'./client/index.js',
-	],
-	output: {
-		path: __dirname,
-		filename: './public/bundle.js',
-	},
-	resolve: {
-		extensions: ['.js', '.jsx'],
-	},
-	devtool: 'source-map',
-	watchOptions: {
-		ignored: /node_modules/,
-	},
-	module: {
-		rules: [
-			{
-				test: /\.jsx?$/,
-				exclude: /node_modules/,
-				loader: 'babel-loader',
-			},
-			{
-				test: /\.css$/,
-				use: ['style-loader', 'css-loader'],
-			},
-			{
-				test: /\.s[ac]ss$/i,
-				use: [
-					'style-loader',
-					'css-loader',
-					{
-						loader: 'sass-loader',
-						options: {
-							// Prefer `dart-sass`
-							implementation: require('sass'),
-							sassOptions: {
-								fiber: require('fibers'),
-							},
-						},
-					},
-				],
-			},
-		],
-	},
+  mode: isDev ? 'development' : 'production',
+  entry: './client/index.js',
+  output: {
+    path: __dirname,
+    filename: './public/bundle.js',
+  },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  },
+  watchOptions: {
+    ignored: /node_modules/,
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader'],
+      },
+      {
+        test: /\.s[ac]ss$/i,
+        use: [
+          'style-loader',
+          'css-loader',
+          {
+            loader: 'sass-loader',
+            options: {
+              // Prefer `dart-sass`
+              implementation: require('sass'),
+            },
+          },
+        ],
+      },
+    ],
+  },
+  devtool: 'eval-cheap-module-source-map',
 }
